[test]add test cases

+++ src/mplfe/jbc_input/include/jbc_opcode.h
@@ -533,7 +533,7 @@ class JBCOpJsr : public JBCOp {
   }
 
   void SetSlotIdx(uint32 argSlotIdx) {
-    slotIdx = argSlotIdx;
+    slotIdx = static_cast<uint16>(argSlotIdx);
   }
 
   int32 GetJsrID() const {
